backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
playwright==1.38.0
prophet==1.1.4
pandas==2.1.0
twilio==8.8.0
pydantic==2.3.0
python-multipart==0.0.6
starlette==0.27.0
typing-extensions==4.7.1